Create test for getShipBlocks()
Before that, I thought that it would be better and safer to keep
ship position private, and when player hit the cell, the cell coords
will be passed to hit() in every ship, which would cause the
hit ship, if any, to take damage.

But eventually I realized that I will need to show player's ship
in the player's board. So, I will need to see where to draw the ship.

I think it's safe as long as ships will be hidden in the Board objects.