Refactor tests for gameboard
Ships in gameboard will be stored just as a string, placed into
a particular group of cells. The ships itself will be accessed via
an array, that contain all the ships.
The reason for that: when a ship is hit and sank, I need to know where
the ship's blocks are to update adjacent cells (mark them 'miss')